UnconvUser: Make both entity and DTO extend BaseUser
BaseUser will implement Spring UserDetails and extending BaseUser will allow me gain access to all the interface methods, avoiding code duplications
