+++ src/main/java/com/unconv/spring/dto/UnconvUserDTO.java
@@ -1,22 +1,20 @@
 package com.unconv.spring.dto;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import java.util.Collection;
+import com.unconv.spring.base.BaseUser;
 import java.util.UUID;
 import javax.validation.constraints.NotEmpty;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
 
 @Getter
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class UnconvUserDTO implements UserDetails {
+public class UnconvUserDTO extends BaseUser {
     private UUID id;
 
     @NotEmpty(message = "Username cannot be empty")
@@ -27,29 +25,4 @@ public class UnconvUserDTO implements UserDetails {
 
     @NotEmpty(message = "Password cannot be empty")
     private String password;
-
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        return null;
-    }
-
-    @Override
-    public boolean isAccountNonExpired() {
-        return false;
-    }
-
-    @Override
-    public boolean isAccountNonLocked() {
-        return false;
-    }
-
-    @Override
-    public boolean isCredentialsNonExpired() {
-        return false;
-    }
-
-    @Override
-    public boolean isEnabled() {
-        return false;
-    }
 }
